pyproject.toml
[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__"]
venv = ".venv"

[tool.flake8]
max-line-length = 50
exclude = [".git", "__pycache__", "venv", ".venv"]
ignore = ["E203", "E266", "E501", "W503"]

[tool.black]
line-length = 79
target-version = ['py39']

[tool.poetry]
name = "data-engineering"
version = "0.1.0"
description = ""
authors = ["Dang Vu <vnhd1995@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^24.11.0"
kafka-python = { git = "https://github.com/dpkp/kafka-python.git" }
schedule = "^1.2.1"
spark = "^0.2.1"
pyspark = "^3.5.1"
jupyter = "^1.0.0"
notebook = "^7.1.3"
isort = "^5.13.2"
flake8 = "^7.0.0"
pandas = "^2.2.2"
requests = "^2.31.0"
flake8-pyproject = "^1.2.3"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
